Split gen tests into separate files
